Created configuration objects for connections
Separation of concerns is performed by having separate classes to handle
the configuration of the connections. Connection's configuration is
defined in PostgresConnection.Configuration interface, and an abstract
class (AbstractPostgresConnectionConfiguration) and a concrete
implementation for TCP/IP connections
(TcpIpPostgresConnectionConfiguration) are provided.

TODO: more refactoring is required in RxPostgresClient to simplify the
Builder making more use of the configuration classes.

Thanks to these classes, clients may now instantiate configuration
objects independently of the connection, and later on set the
configuration to the connection by having code similar to:

PostgresConnection.Configuration config = new
TcpIpPostgresConnectionConfiguration("host", 5432);

RxPostgresClient client = RxPostgresClient
                .newConfiguration()
				...
                .configuration(config)
				...